+++ projects/rocfft/CHANGELOG.md
@@ -14,7 +14,7 @@ Documentation for rocFFT is available at
 
 * Fixed potential division by zero when constructing plans using dimensions of length 1.
 * Fixed result scaling on multi-device transforms.
-* Fixed callbacks on single-process multi-GPU transforms.
+* Fixed callbacks on multi-device transforms.
 * Fixed potential issue with data generation for multidimensional transforms in rocfft-tests and rocfft-bench.
 * Fixed issue that sometimes blocked complex-to-complex FFT plan creation when using noncontiguous strides in multiple dimensions.
 * Fixed issue that sometimes blocked complex-to-real FFT plan creation when using noncontiguous strides with small lengths on the two fastest dimensions.